Thought a bit about the file transfer process, which is roughly outlined as follows:
- Queue a file for download;
- Once it pops off the queue, a DownloadTransfer object is created;
- This DownloadTransfer object fires off a TTH search for alternate sources;
- Once sources are obtained, it requests the TTH tree of the file it is about to download from one of the sources;
- Once the TTH tree is obtained, the actual file transfer can start.

We need signal/slot pairs to match TTH searches, much like queryTTH, but just giving a yes/no type answer. In signal/slot lingo it would mean it would trigger the response slot only if a file matching the TTH root is shared. TTH searches are simple in nature, it just pings the network for hosts sharing a specific file without requesting additional information about it.

We also need signal/slot pairs to handle the requests and responses of TTH Tree questions. The format of these responses is still up for debate. I suggest spanning the TTH tree over several datagrams if necessary, beginning each one with the root TTH, following with (quint32 bucket number)(QByteArray bucketTTH)etc...

There is a chance that some of the packets might get lost, so we need parameters for ranged queries on the TTH tree, not currently implemented in the prototype interfaces. Just being philosophical here.

I commented the relevant section in arpmanetdc.cpp, which is up for comment :)