worked on transfers, among other: - fixed an infinite loop in transfermanager - fixed protocol parsing in upload request path - fixed hash tree reply in sharesearch to prepend tth to packets
transfers work up to the point where the hash tree is downloaded and the actual transfer must begin. it is currently stuck where the upload is queued at the remote client and the file name of the tth is requested.

tree downloading fails for very large files. the iso image of fedora 15 should have over 3400 buckets, but it stops after 2146 buckets are retrieved in order. the anomaly is probably not due to packet loss, because the packets arrive in order and subsequent retries yield the same result.

work in progress...