Transfers are transferring *something* now, I do not yet know it is the real thing, but on the wire it looks like binary data when I transfer a 10MB iso file.
Added finished detection, quite necessary...

We need a method to notify the GUI when a download has finished, so that it can be taken off the queue. I shall shortly implement a signal for it on this side.

Another issue I have seen is that it seems as if I am unable to unshare files in Linux. At first I thought the search searches files flagged as inactive, but on second inspection it seems as if the actual inactive flagging does not happen.