forgot to open input file in uploadtransfer after reshuffling some things. fixed.
also temporarily short circuited the bucket hash signals to directly flush the buckets to disk to see what is inside. a quick sha1sum confirmed that the 10.4 MB of data I transferred is INDEED INTACT! also rather quick for such a primitive protocol, it takes between 1 and 2 seconds, so I guess the transfer rate sits somewhere between 5 and 10 MB/s!