Found the activeSegments miscalculation
- Added +1 before knowing the segment creation succeeded
- Also, the new segment timer was never started

Actually, it seems as though it's working. Still, there seems to be
clients that respond to a search request but never to a download request.
The segment fails correctly and is recreated, but never receive any data.
I'm guessing this might have to do with broken/moved files - although
these should be corrected every hour when shares are updated...